Improve plain text editor
- unwrap paragraphs when switching to rich text
- hard wrap when switching to plain text
- stop bullet list on double enter
